# FamilyBenefits-api-openapi.yaml
openapi: 3.0.3
info:
  title: Family Benefits
  description: rest-api для приложения "Пособия для семьи с детьми"
  version: 1.0.0
  contact:
    email: smegovic@gmail.com
servers:
  - url: https://localhost
...
paths:
...
  /user:
    post:
      tags:
        - user
      summary: Добавляет нового пользователя. Регистрация гостя
      description:
        Добавляет нового пользователя. Регистрация гостя.
        Для анонимного клиента.
      operationId: addUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAdd'
        required: true
      responses:
        '201':
          description: Пользователь зарегистрирован
          content: { }
        '400':
          description:
            Варианты -
            1) администратор или пользователь с данным email уже существует;
            2) не безопасный пароль;
            3) указанные пароли не эквивалентны;
            4) указанный "email" не является email;
            5) даты рождения пользователя или детей не соответствуют формату "dd.mm.yyyy";
            6) даты рождения пользователя или детей позже текущей даты.
        '404':
          description: Город или критерии с указанными ID не найдены
          content: { }
      security: [ ]
    put:
      tags:
        - user
      summary: Обновляет данные пользователя
      description:
        Обновляет данные пользователя.
        О запросившем клиенте.
        Для пользователя.
        Необходима авторизация.
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '201':
          description: Пользователь обновлен
          content: { }
        '400':
          description:
            Варианты -
            1) указанный "email" не является email;
            2) даты рождения пользователя или детей не соответствуют формату "dd.mm.yyyy";
            3) даты рождения пользователя или детей позже текущей даты.
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description:
            Варианты -
            1) пользователь с указанными данными не найден;
            2) город с указанным ID не найден.
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
  /user/{id}:
    delete:
      tags:
        - user
      summary: Удаляет пользователя по ID
      description:
        Удаляет пользователя по ID. Если пользователь имеет роль админа, то удаляются роль пользователя и пользовательские данные.
        О запросившем клиенте.
        Для пользователя.
        Необходима авторизация.
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: ID пользователя
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '201':
          description: Пользователь удален
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description: Пользователь с указанными данными не найден
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
    get:
      tags:
        - user
      summary: Возвращает информацию о пользователе
      description:
        Возвращает информацию о пользователе по ID.
        О запросившем клиенте.
        Для пользователя.
        Необходима авторизация.
      operationId: getUser
      parameters:
        - name: id
          in: path
          description: ID пользователя
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '200':
          description: Данные о пользователе получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description: Пользователь с указанными данными не найден
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
  /user/{id}/benefits:
    get:
      tags:
        - user
      summary: Возвращает множество предложенных пособий пользователя
      description:
        Возвращает множество предложенных пособий.
        Для пользователя.
        Необходима авторизация.
      operationId: getUserBenefits
      parameters:
        - name: id
          in: path
          description: ID пользователя
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '200':
          description: Множество предложенных пособий получено
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BenefitInfo'
        '400':
          description: Критерии пользователя устарели
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description: Не найден пользователь с указанными данными
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserInfo:
      description: Информация о пользователе
      properties:
        id:
          description: ID пользователя
          type: string
          example: td56387fhj
        name:
          description: Имя пользователя
          type: string
          example: name
        email:
          description: Электронная почта пользователя
          type: string
          example: exmaple@mail.com
        isVerifiedEmail:
          description: Статус почты, true, если подтверждена
          type: boolean
          example: false
        birthDateChildren:
          description: Множество дат рождений детей пользователя
          type: array
          items:
            type: string
            format: date
            example: 01.01.2022
        dateBirth:
          description: Дата рождения пользователя
          type: string
          format: date
          example: 01.01.2022
        nameRoleSet:
          description: Множество названий ролей пользователя
          type: array
          items:
            description: Название роли
            type: string
            example: name
        shortCity:
          $ref: '#/components/schemas/ObjectShortInfo'
        criterionSet:
          description: Множество информаций о критериях пользователя
          type: array
          items:
            $ref: '#/components/schemas/CriterionInfo'
    UserAdd:
      description: Объект запроса для добавления пользователя
      properties:
        name:
          description: Имя пользователя
          type: string
          example: name
        email:
          description: Электронная почта пользователя
          type: string
          example: exmaple@mail.com
        password:
          description: Пароль пользователя
          type: string
          example: password
        repeatPassword:
          description: Повторно введенный пароль пользователя
          type: string
          example: password
        birthDateChildren:
          description: Множество дат рождений детей пользователя
          type: array
          items:
            description: Дата рождения ребенка
            type: string
            format: date
            example: 01.01.2022
        dateBirth:
          description: Дата рождения пользователя
          type: string
          format: date
          example: 01.01.2022
        idCity:
          description: ID города пользователя
          type: string
          example: td56387fhj
        idCriterionSet:
          description: Множество ID критерий пользователя
          type: array
          items:
            description: ID критерия
            type: string
            example: td56387fhj
    UserUpdate:
      description: Объект запроса для обновления пользователя
      properties:
        id:
          description: ID пользователя
          type: string
          example: td56387fhj
        name:
          description: Имя пользователя
          type: string
          example: name
        email:
          description: Электронная почта пользователя
          type: string
          example: exmaple@mail.com
        birthDateChildren:
          description: Множество дат рождений детей пользователя
          type: array
          items:
            description: Дата рождения ребенка
            type: string
            format: date
            example: 01.01.2022
        dateBirth:
          description: Дата рождения пользователя
          type: string
          format: date
          example: 01.01.2022
        idCity:
          description: ID города пользователя
          type: string
          example: td56387fhj
        idCriterionSet:
          description: Множество ID критерий пользователя
          type: array
          items:
            description: ID критерия
            type: string
            example: td56387fhj
...
    CriterionInfo:
      description: Информация о критерии для подбора пособий
      properties:
        id:
          description: ID критерия
          type: string
          example: td56387fhj
        name:
          description: Название критерия
          type: string
          example: name
        info:
          description: Информация критерия
          type: string
          example: info
        shortCriterionType:
          $ref: '#/components/schemas/ObjectShortInfo'
        shortBenefitSet:
          description: Множество кратких информаций о пособиях критерия
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
...
    BenefitInfo:
      description: Информация о пособии
      properties:
        id:
          description: ID пособия
          type: string
          example: td56387fhj
        name:
          description: Название пособия
          type: string
          example: name
        info:
          description: Информация пособия
          type: string
          example: info
        documents:
          description: Документы для получения пособия
          type: string
          example: documents
        shortCitySet:
          description: Множество кратких информаций о городах пособия
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
        shortInstitutionSet:
          description: Множество кратких информаций об учреждениях пособия
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
        criterionSet:
          description: Множество информаций о критериях пособия
          type: array
          items:
            $ref: '#/components/schemas/CriterionInfo'
...
    ObjectShortInfo:
      description: Краткая информация об объекте
      properties:
        idObject:
          description: ID объекта
          type: string
          example: td56387fhj
        nameObject:
          description: Название объекта
          type: string
          example: name